Refactor Editor into its own component
