+++ dashboard/src/components/Editor.js
@@ -1,64 +1,54 @@
 import React, { Component } from "react";
-import "../assets/css/App.css";
 
-import "brace";
-import AceEditor from "react-ace";
-import "brace/mode/logiql";
-import "brace/theme/github";
+import { timeout, checkStatus, parseJSON } from "./Helpers";
+import "../assets/css/Editor.css";
+
+require("codemirror/addon/hint/show-hint.css");
+
+function isNotEmpty(response) {
+  let keys = Object.keys(response);
+  for (let i = 0; i < keys.length; i++) {
+    if (keys[i] !== "server_latency" && keys[i] !== "uids") {
+      return keys[i].length > 0;
+    }
+  }
+  return false;
+}
+
+function handleResponse(result) {
+  // This is the case in which user sends a mutation. We display the response from server.
+  if (result.code !== undefined && result.message !== undefined) {
+    this.props.storeQuery(this.state.query);
+    this.props.renderResText("success-res", JSON.stringify(result, null, 2));
+  } else if (isNotEmpty(result)) {
+    this.props.storeQuery(this.state.query);
+    this.props.renderGraph(result);
+  } else {
+    // We probably didn't get any results.
+    this.props.renderResText(
+      "error-res",
+      "Your query did not return any results.",
+    );
+  }
+}
 
 class Editor extends Component {
-  lastQuery = () => {
-    var queries = JSON.parse(localStorage.getItem("queries"));
-    if (queries == null) {
-      return [undefined, "", []];
-    }
-    // This means queries has atleast one element.
+  constructor(props: Props) {
+    super(props);
+    this.state = {
+      query: this.props.query,
+    };
+  }
 
-    return [0, queries[0], queries];
-  };
-
-  nextQuery = () => {
-    var queries = JSON.parse(localStorage.getItem("queries"));
-    if (queries == null) {
-      return;
-    }
-
-    var idx = this.state.queryIndex;
-    if (idx === undefined || idx - 1 < 0) {
-      return;
-    }
-    this.setState({
-      query: queries[idx - 1],
-      queryIndex: idx - 1,
-    });
-  };
-
-  previousQuery = () => {
-    var queries = JSON.parse(localStorage.getItem("queries"));
-    if (queries == null) {
-      return;
-    }
-
-    var idx = this.state.queryIndex;
-    if (idx === undefined || idx + 1 === queries.length) {
-      return;
-    }
-    this.setState({
-      queryIndex: idx + 1,
-      query: queries[idx + 1],
-    });
-  };
-
-  runQuery = e => {
+  runQuery = (e: Event) => {
     e.preventDefault();
-    // Resetting state
-    network && network.destroy();
-    this.setState(this.resetState());
+    // Resetting state of parent related to a query.
+    this.props.resetState();
 
     var that = this;
     timeout(
       60000,
-      fetch("http://localhost:8080/query?debug=true", {
+      fetch(process.env.REACT_APP_DGRAPH + "/query?debug=true", {
         method: "POST",
         mode: "cors",
         headers: {
@@ -68,67 +58,7 @@ class Editor extends Component {
       })
         .then(checkStatus)
         .then(parseJSON)
-        .then(function(result) {
-          var key = Object.keys(result)[0];
-          if (result.code !== undefined && result.message !== undefined) {
-            that.storeQuery(that.state.query);
-            // This is the case in which user sends a mutation. We display the response from server.
-            that.setState({
-              resType: "success-res",
-              response: JSON.stringify(result, null, 2),
-            });
-          } else if (key !== undefined) {
-            that.storeQuery(that.state.query);
-            // We got the result for a query.
-            var response = result[key];
-            that.setState({
-              latency: result.server_latency.total,
-            });
-            var startTime = new Date();
-
-            setTimeout(
-              function() {
-                var graph = processGraph(response, key, -1);
-                nodeSet = new vis.DataSet(graph[0]);
-                edgeSet = new vis.DataSet(graph[1]);
-                that.setState({
-                  nodes: graph[0].length,
-                  relations: graph[1].length,
-                });
-              },
-              1000,
-            );
-
-            // We call procesGraph with a 40 node limit and calculate the whole dataset in
-            // the background.
-            var graph = processGraph(response, key, 20);
-            setTimeout(
-              (function() {
-                that.setState({
-                  partial: this.state.nodes > graph[0].length,
-                });
-              }).bind(that),
-              1000,
-            );
-
-            renderNetwork.call(that, graph[0], graph[1]);
-
-            var endTime = new Date();
-            var timeTaken = ((endTime.getTime() - startTime.getTime()) /
-              1000).toFixed(2);
-            that.setState({
-              rendering: timeTaken + "s",
-              resType: "success-res",
-            });
-          } else {
-            console.warn("We shouldn't be here really");
-            // We probably didn't get any results.
-            that.setState({
-              resType: "error-res",
-              response: "Your query did not return any results.",
-            });
-          }
-        }),
+        .then(handleResponse.bind(that)),
     )
       .catch(function(error) {
         console.log(error.stack);
@@ -137,67 +67,207 @@ class Editor extends Component {
       })
       .then(function(errorMsg) {
         if (errorMsg !== undefined) {
-          that.setState({
-            response: errorMsg,
-            resType: "error-res",
-          });
+          this.props.renderResText("error-res", errorMsg);
         }
       });
   };
 
   render() {
     return (
-      <div className="App-editor">
-        <AceEditor
-          mode="logiql"
-          theme="github"
-          name="editor"
-          editorProps={{ $blockScrolling: Infinity }}
-          width="100%"
-          height="350px"
-          fontSize="12px"
-          ref="code"
-          value={this.props.query}
-          focus={true}
-          showPrintMargin={false}
-          wrapEnabled={true}
-          onChange={this.propsqueryChange}
-          commands={[
-            {
-              name: "runQuery",
-              bindKey: {
-                win: "Ctrl-Enter",
-                mac: "Command-Enter",
-              },
-              exec: (function(editor) {
-                this.runQuery(new Event(""));
-              }).bind(this),
-            },
-            {
-              name: "previousQuery",
-              bindKey: {
-                win: "Ctrl-Up",
-                mac: "Command-Up",
-              },
-              exec: (function(editor) {
-                this.previousQuery();
-              }).bind(this),
-            },
-            {
-              name: "nextQuery",
-              bindKey: {
-                win: "Ctrl-Down",
-                mac: "Command-Down",
-              },
-              exec: (function(editor) {
-                this.nextQuery();
-              }).bind(this),
-            },
-          ]}
+      <div>
+        <form id="query-form">
+          <div className="form-group">
+            <label htmlFor="query">Query</label>
+            <button
+              type="submit"
+              className="btn btn-primary pull-right"
+              onClick={this.runQuery}
+            >
+              Run
+            </button>
+          </div>
+        </form>
+
+        <div
+          className="Editor-basic"
+          ref={editor => {
+            this._editor = editor;
+          }}
         />
       </div>
     );
   }
+
+  componentWillReceiveProps = nextProps => {
+    if (nextProps.query !== undefined && this.state.query !== nextProps.query) {
+      this.editor.setValue(nextProps.query);
+    }
+  };
+
+  componentDidMount = () => {
+    const CodeMirror = require("codemirror");
+    require("codemirror/addon/hint/show-hint");
+    require("codemirror/addon/comment/comment");
+    require("codemirror/addon/edit/matchbrackets");
+    require("codemirror/addon/edit/closebrackets");
+    require("codemirror/addon/fold/foldcode");
+    require("codemirror/addon/fold/foldgutter");
+    require("codemirror/addon/fold/brace-fold");
+    require("codemirror/addon/lint/lint");
+    require("codemirror/keymap/sublime");
+    require("codemirror-graphql/hint");
+    require("codemirror-graphql/lint");
+    require("codemirror-graphql/info");
+    require("codemirror-graphql/jump");
+    require("codemirror-graphql/mode");
+
+    let keywords = [];
+    timeout(
+      1000,
+      fetch(process.env.REACT_APP_DGRAPH + "/ui/keywords", {
+        method: "GET",
+        mode: "cors",
+      })
+        .then(checkStatus)
+        .then(parseJSON)
+        .then(function(result) {
+          keywords = result.keywords.map(function(kw) {
+            return kw.name;
+          });
+        }),
+    )
+      .catch(function(error) {
+        console.log(error.stack);
+        console.warn(
+          "In catch: Error while trying to fetch list of keywords",
+          error,
+        );
+        return error;
+      })
+      .then(function(errorMsg) {
+        if (errorMsg !== undefined) {
+          console.warn(
+            "Error while trying to fetch list of keywords",
+            errorMsg,
+          );
+        }
+      });
+
+    this.editor = CodeMirror(this._editor, {
+      value: this.props.query,
+      lineNumbers: true,
+      tabSize: 2,
+      lineWrapping: true,
+      mode: "graphql",
+      theme: "graphiql",
+      keyMap: "sublime",
+      autoCloseBrackets: true,
+      completeSingle: false,
+      showCursorWhenSelecting: true,
+      foldGutter: true,
+      gutters: ["CodeMirror-linenumbers", "CodeMirror-foldgutter"],
+      extraKeys: {
+        "Ctrl-Space": cm => {
+          CodeMirror.commands.autocomplete(cm);
+        },
+        "Cmd-Space": cm => {
+          CodeMirror.commands.autocomplete(cm);
+        },
+        "Cmd-Enter": () => {
+          this.runQuery(new Event(""));
+        },
+        "Ctrl-Enter": () => {
+          this.runQuery(new Event(""));
+        },
+      },
+      autofocus: true,
+    });
+
+    this.editor.setCursor(this.editor.lineCount(), 0);
+
+    this.editor.on(
+      "change",
+      (function(cm, obj) {
+        this.setState({
+          query: obj.text.join("\n"),
+        });
+      }).bind(this),
+    );
+
+    CodeMirror.registerHelper("hint", "fromList", function(cm, options) {
+      var cur = cm.getCursor(), token = cm.getTokenAt(cur);
+
+      // This is so that we automatically have a space before (, so that auto-
+      // complete inside braces works. Otherwise it doesn't work for
+      // director.film(orderasc: release_date).
+      let openBrac = token.string.indexOf("(");
+      if (
+        openBrac !== -1 &&
+        token.string[openBrac - 1] != undefined &&
+        token.string[openBrac - 1] != " "
+      ) {
+        let oldString = token.string.substr(openBrac);
+        cm.replaceRange(" ", { line: cur.line, ch: token.start + openBrac });
+      }
+
+      var to = CodeMirror.Pos(cur.line, token.end);
+      if (token.string) {
+        var term = token.string, from = CodeMirror.Pos(cur.line, token.start);
+      } else {
+        var term = "", from = to;
+      }
+
+      // So that we don't autosuggest for anyof/allof filter values which
+      // would be inside quotes.
+      if (term.length > 0 && term[0] === '"') {
+        return { list: [], from: from, to: to };
+      }
+
+      // TODO - This is a hack because Graphiql mode considers . as an invalidchar.
+      // Ideally we should write our own mode which allows . in predicate.
+      if (
+        token.type === "invalidchar" &&
+        token.state.prevState !== undefined &&
+        token.state.prevState.kind === "Field"
+      ) {
+        term = token.state.prevState.name + token.string;
+        from.ch = from.ch - token.state.prevState.name.length;
+      }
+
+      // because Codemirror strips the @ from a directive.
+      if (token.state.kind === "Directive") {
+        term = "@" + term;
+        from.ch -= 1;
+      }
+
+      term = term.toLowerCase();
+
+      var found = [];
+      for (var i = 0; i < options.words.length; i++) {
+        var word = options.words[i];
+        if (term.length > 0 && word.startsWith(term)) {
+          found.push(word);
+        }
+      }
+
+      if (found.length) return { list: found, from: from, to: to };
+    });
+
+    CodeMirror.commands.autocomplete = function(cm) {
+      CodeMirror.showHint(cm, CodeMirror.hint.fromList, {
+        completeSingle: false,
+        words: keywords,
+      });
+    };
+
+    this.editor.on("keydown", function(cm, event) {
+      const code = event.keyCode;
+
+      if (!event.ctrlKey && code >= 65 && code <= 90) {
+        CodeMirror.commands.autocomplete(cm);
+      }
+    });
+  };
 }
 
 export default Editor;
